PDFBOX-1958: use int instead of float, remove TODO
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1780089 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/graphics/image/SampledImageReader.java
@@ -81,8 +81,8 @@ final class SampledImageReader
         WritableRaster raster = masked.getRaster();
         WritableRaster alpha = mask.getRaster();
 
-        final float[] transparent = new float[4];
-        float[] alphaPixel = null;
+        final int[] transparent = new int[4];
+        int[] alphaPixel = null;
         for (int y = 0; y < height; y++)
         {
             for (int x = 0; x < width; x++)
